fix: resolve circular dependency in crc32c
